requirements.txt
celery
django
django-celery
django-tastypie
feedparser
gunicorn
opml
psycopg2
requests
git+ssh://git@github.com/AmeliaKnows/libgreader.git#egg=libgreader
google-api-python-client
